Split direct build graph execution tests
